create new file to l10n_br_account_product_cfop
