+++ l10n_br_account_product/models/__init__.py
@@ -8,6 +8,7 @@ from . import l10n_br_account_fiscal_category
 from . import l10n_br_account_document_serie
 from . import l10n_br_account_partner_fiscal
 from . import l10n_br_account_product
+from . import l10n_br_account_product_cfop
 from . import res_company
 from . import l10n_br_tax_definition_company_product
 from . import l10n_br_account
